#3866 迁移组件装配逻辑从 data 层到 trading 层
- EngineAssemblyService 新增 assemble_live_portfolio() 方法
- Data 层 portfolio_service 删除跨层代码（load_portfolio_with_components 及 4 个辅助方法）
- 3 个调用方（TaskEngineBuilder、ExecutionNode、BacktestTaskProcessor）改为调 trading 层
- build_position_writer/build_redis_writer 从私有方法改为公开，供 trading 层调用
- Data 层不再有任何 from ginkgo.trading 导入
- 新增 5 个 TDD 测试覆盖 assemble_live_portfolio 行为
