+++ src/ginkgo/trading/services/engine_assembly_service.py
@@ -24,6 +24,7 @@ from typing import Optional, Dict, Any, List, Union
 from pathlib import Path
 from decimal import Decimal
 
+import pandas as pd
 from ginkgo.libs import GLOG, GinkgoLogger
 from ginkgo.enums import EVENT_TYPES
 from ginkgo.trading.engines import BaseEngine, BacktestEngine
@@ -136,6 +137,305 @@ class EngineAssemblyService(BaseService):
         else:
             self._logger.WARN(f"⚠️ Component {component.__class__.__name__} doesn't support ID injection")
             return False
+
+    # ========== Live Portfolio 装配 ==========
+
+    def assemble_live_portfolio(
+        self,
+        portfolio_id: str,
+        position_writer=None,
+        redis_writer=None,
+    ) -> ServiceResult:
+        """
+        从数据库配置装配完整的 PortfolioLive 实例。
+
+        Data 层通过 position_writer/redis_writer 提供持久化基础设施，
+        本方法只负责 trading 层的组件装配逻辑。
+
+        Args:
+            portfolio_id: Portfolio UUID
+            position_writer: Data 层的持仓持久化服务
+            redis_writer: Data 层的 Redis 状态写入服务
+
+        Returns:
+            ServiceResult: 包含完整 PortfolioLive 实例
+        """
+        try:
+            # 1. 从 Data 层获取 Portfolio 原始数据
+            portfolio_result = self._portfolio_service.get(portfolio_id=portfolio_id)
+            if not portfolio_result.is_success():
+                return ServiceResult(success=False, error=f"Portfolio不存在: {portfolio_id}")
+
+            portfolio_model = portfolio_result.data
+            if isinstance(portfolio_model, pd.DataFrame):
+                if portfolio_model.shape[0] == 0:
+                    return ServiceResult(success=False, error=f"Portfolio不存在: {portfolio_id}")
+                portfolio_model = portfolio_model.iloc[0]
+            elif hasattr(portfolio_model, '__iter__') and not isinstance(portfolio_model, dict):
+                items = list(portfolio_model)
+                portfolio_model = items[0] if len(items) > 0 else None
+                if portfolio_model is None:
+                    return ServiceResult(success=False, error=f"Portfolio不存在: {portfolio_id}")
+
+            # 2. 创建 PortfolioLive 实例
+            from ginkgo.trading.portfolios.portfolio_live import PortfolioLive
+
+            portfolio = PortfolioLive(
+                uuid=portfolio_model.uuid,
+                name=portfolio_model.name,
+                position_writer=position_writer,
+                redis_writer=redis_writer,
+            )
+
+            portfolio.add_cash(float(portfolio_model.initial_capital))
+            portfolio.initial_capital = float(portfolio_model.initial_capital)
+            if not hasattr(portfolio, 'frozen_cash'):
+                portfolio.frozen_cash = 0.0
+
+            # 3. 设置上下文
+            from ginkgo.trading.context.engine_context import EngineContext
+            from ginkgo.trading.context.portfolio_context import PortfolioContext
+
+            engine_context = EngineContext(engine_id="livecore")
+            engine_context.set_task_id(portfolio.uuid)
+            portfolio_context = PortfolioContext(
+                portfolio_id=portfolio.uuid,
+                engine_context=engine_context,
+            )
+            portfolio._context = portfolio_context
+
+            # 4. 获取并加载组件
+            components_result = self._portfolio_service.get_components(portfolio_id=portfolio_id)
+            if not components_result.is_success() or not components_result.data:
+                self._bind_default_components(portfolio)
+            else:
+                components = components_result.data
+                if not components or len(components) == 0:
+                    self._bind_default_components(portfolio)
+                else:
+                    self._bind_components_from_config(portfolio, components)
+
+            self._logger.INFO(f"✓ Portfolio {portfolio_id[:8]}... 加载成功，包含所有组件")
+            return ServiceResult(success=True, data=portfolio)
+
+        except Exception as e:
+            self._logger.ERROR(f"加载Portfolio失败: {str(e)}")
+            import traceback
+            self._logger.ERROR(f"Traceback: {traceback.format_exc()}")
+            return ServiceResult(success=False, error=f"加载Portfolio失败: {str(e)}")
+
+    def _bind_default_components(self, portfolio):
+        """绑定默认组件到Portfolio"""
+        from ginkgo.trading.strategies.random_signal_strategy import RandomSignalStrategy
+        from ginkgo.trading.selectors.cn_all_selector import CNAllSelector
+        from ginkgo.trading.sizers.fixed_sizer import FixedSizer
+        from ginkgo.trading.risk_management.position_ratio_risk import PositionRatioRisk
+
+        portfolio.add_strategy(RandomSignalStrategy())
+        portfolio.bind_selector(CNAllSelector())
+        portfolio.bind_sizer(FixedSizer())
+        portfolio.add_risk_manager(PositionRatioRisk())
+
+    def _bind_components_from_config(self, portfolio, components: list):
+        """从组件配置实例化并绑定到Portfolio"""
+        from ginkgo.enums import FILE_TYPES
+
+        strategies_list = []
+        selectors_list = []
+        sizers_list = []
+        risk_managers_list = []
+
+        for component in components:
+            comp_type = component.get('component_type')
+            if isinstance(comp_type, str):
+                if comp_type.isdigit():
+                    comp_type_int = int(comp_type)
+                else:
+                    try:
+                        comp_type_int = FILE_TYPES[comp_type].value
+                    except Exception:
+                        continue
+            else:
+                comp_type_int = int(comp_type)
+
+            if comp_type_int == FILE_TYPES.STRATEGY.value:
+                strategies_list.append(component)
+            elif comp_type_int == FILE_TYPES.SELECTOR.value:
+                selectors_list.append(component)
+            elif comp_type_int == FILE_TYPES.SIZER.value:
+                sizers_list.append(component)
+            elif comp_type_int == FILE_TYPES.RISKMANAGER.value:
+                risk_managers_list.append(component)
+
+        # 加载策略
+        if len(strategies_list) == 0:
+            from ginkgo.trading.strategies.random_signal_strategy import RandomSignalStrategy
+            portfolio.add_strategy(RandomSignalStrategy())
+        else:
+            for strategy_config in strategies_list:
+                strategy = self._instantiate_component(strategy_config, 'strategy')
+                if strategy:
+                    portfolio.add_strategy(strategy)
+
+        # 加载选股器
+        if len(selectors_list) == 0:
+            from ginkgo.trading.selectors.cn_all_selector import CNAllSelector
+            portfolio.bind_selector(CNAllSelector())
+        else:
+            selector_config = selectors_list[0]
+            selector = self._instantiate_component(selector_config, 'selector')
+            if selector:
+                portfolio.bind_selector(selector)
+
+        # 加载Sizer
+        if len(sizers_list) == 0:
+            from ginkgo.trading.sizers.fixed_sizer import FixedSizer
+            portfolio.bind_sizer(FixedSizer())
+        else:
+            sizer_config = sizers_list[0]
+            sizer = self._instantiate_component(sizer_config, 'sizer')
+            if sizer:
+                portfolio.bind_sizer(sizer)
+
+        # 加载风控
+        if len(risk_managers_list) == 0:
+            from ginkgo.trading.risk_management.position_ratio_risk import PositionRatioRisk
+            portfolio.add_risk_manager(PositionRatioRisk())
+        else:
+            for risk_config in risk_managers_list:
+                risk_manager = self._instantiate_component(risk_config, 'risk_management')
+                if risk_manager:
+                    portfolio.add_risk_manager(risk_manager)
+
+    def _instantiate_component(self, component_config: dict, component_type: str):
+        """从组件配置实例化组件"""
+        try:
+            file_id = component_config.get('component_id')
+            if not file_id:
+                return self._get_default_component(component_type)
+
+            mount_id = component_config.get('mount_id')
+            component_params = []
+
+            if mount_id:
+                try:
+                    from ginkgo.data.containers import container
+                    param_crud = container.cruds.param()
+                    param_records = param_crud.find(filters={"mapping_id": mount_id})
+                    if param_records:
+                        sorted_params = sorted(param_records, key=lambda p: p.index)
+                        component_params = [param.value for param in sorted_params]
+                except Exception as e:
+                    self._logger.WARN(f"获取参数失败: {e}")
+
+            from ginkgo import services
+            file_service = services.data.file_service()
+            file_result = file_service.get_by_uuid(file_id)
+
+            if not file_result.success or not file_result.data:
+                return self._get_default_component(component_type)
+
+            file_info = file_result.data
+            if isinstance(file_info, dict) and "file" in file_info:
+                mfile = file_info["file"]
+                if hasattr(mfile, "data") and mfile.data:
+                    code_content = mfile.data.decode("utf-8", errors="ignore") if isinstance(mfile.data, bytes) else str(mfile.data)
+                else:
+                    return self._get_default_component(component_type)
+            else:
+                return self._get_default_component(component_type)
+
+            import tempfile
+            import importlib.util
+            import os
+
+            with tempfile.NamedTemporaryFile(mode="w", suffix=".py", delete=False) as temp_file:
+                temp_file.write(code_content)
+                temp_file_path = temp_file.name
+
+            try:
+                spec = importlib.util.spec_from_file_location("dynamic_component", temp_file_path)
+                module = importlib.util.module_from_spec(spec)
+                spec.loader.exec_module(module)
+
+                component_class = None
+                for attr_name in dir(module):
+                    if attr_name.startswith("_"):
+                        continue
+                    attr = getattr(module, attr_name)
+                    if isinstance(attr, type) and hasattr(attr, "__bases__"):
+                        is_component = False
+                        if hasattr(attr, "__abstract__") and not getattr(attr, "__abstract__", True):
+                            is_component = True
+                        else:
+                            for base in attr.__bases__:
+                                base_name = base.__name__
+                                if base_name.endswith(("Strategy", "Selector", "Sizer", "RiskManagement")) or \
+                                   base_name in ("BaseStrategy", "BaseSelector", "BaseSizer", "BaseRiskManagement"):
+                                    is_component = True
+                                    break
+                        if is_component:
+                            component_class = attr
+                            break
+
+                if component_class is None:
+                    return self._get_default_component(component_type)
+
+                import inspect
+                if component_params:
+                    converted = self._convert_params(component_class, component_params)
+                    return component_class(*converted)
+                return component_class()
+
+            finally:
+                try:
+                    os.unlink(temp_file_path)
+                except Exception:
+                    pass
+
+        except Exception as e:
+            self._logger.ERROR(f"实例化组件失败: {e}")
+            return self._get_default_component(component_type)
+
+    @staticmethod
+    def _convert_params(component_class, raw_params: list) -> list:
+        """根据构造函数类型注解自动转换参数类型"""
+        import inspect
+        from ginkgo.libs.data.number import convert_to_float, convert_to_int
+        try:
+            sig = inspect.signature(component_class.__init__)
+            params = list(sig.parameters.values())[1:]
+            converted = []
+            for i, raw in enumerate(raw_params):
+                if i < len(params):
+                    ann = params[i].annotation
+                    if ann is int:
+                        converted.append(convert_to_int(raw))
+                    elif ann is float:
+                        converted.append(convert_to_float(raw))
+                    else:
+                        converted.append(raw)
+                else:
+                    converted.append(raw)
+            return converted
+        except Exception:
+            return raw_params
+
+    def _get_default_component(self, component_type: str):
+        """获取默认组件"""
+        from ginkgo.trading.strategies.random_signal_strategy import RandomSignalStrategy
+        from ginkgo.trading.selectors.cn_all_selector import CNAllSelector
+        from ginkgo.trading.sizers.fixed_sizer import FixedSizer
+        from ginkgo.trading.risk_management.position_ratio_risk import PositionRatioRisk
+
+        defaults = {
+            'strategy': RandomSignalStrategy,
+            'selector': CNAllSelector,
+            'sizer': FixedSizer,
+            'risk_management': PositionRatioRisk,
+        }
+        cls = defaults.get(component_type)
+        return cls() if cls else None
 
     # ========== 公共 API ==========
